+++ NEWS.md
@@ -1,3 +1,9 @@
+# ClustImpute 0.1.4
+
+* There is now a new option assign_with_wf. If set to False, then the weight function is only applied in the centroid computation, but ignored in the cluster assignment
+* New check for duplicate centroids. Duplicate rows are replaced with random draws.
+* Compatibility with new version 1.2.2 of ClusterR
+
 # ClustImpute 0.1.3
 
 * Full test coverage